# docs/.generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BIYA DEX API（网关聚合）",
    "version": "7e01d0910fb0aab261873530147fcf5ec1e4776a",
...
  },
...
  "paths": {
    "/exchange": {
      "post": {
        "tags": [
          "Exchange"
        ],
        "summary": "提交已签名的 exchange action",
        "description": "同步完成本地校验（签名标准化、参数校验、签名恢复验签）和入池后返回。accepted 不代表链上执行结果；重复提交返回 1002。",
        "operationId": "postExchange",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Exchange__SignedAction"
              },
              "examples": {
                "limitOrder": {
                  "value": {
                    "action": {
                      "type": "order",
                      "orders": [
                        {
                          "a": 0,
                          "b": true,
                          "p": "65000",
                          "s": "0.001",
                          "r": false,
                          "t": {
                            "limit": {
                              "tif": "Gtc"
                            }
                          }
                        }
                      ],
                      "grouping": "na"
                    },
                    "nonce": 1774952773999,
                    "signature": {
                      "r": "0x<32-byte-r>",
                      "s": "0x<32-byte-s>",
                      "v": 27
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "本地处理结果（注意：业务失败也走 200 + code 区分，见返回码表）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exchange__ExchangeResult"
                }
              }
            }
          },
          "400": {
            "description": "malformed JSON 或字段类型错误，Axum 在进业务逻辑前拒绝。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exchange__Error"
                }
              }
            }
          }
        }
      }
    },
    "/info": {
      "post": {
        "tags": [
          "Exchange",
          "State",
          "Indexer"
        ],
        "summary": "统一查询入口（网关按 type 分流）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Exchange__InfoRequest"
                  },
                  {
                    "$ref": "#/components/schemas/State__InfoRequest"
                  },
                  {
                    "$ref": "#/components/schemas/Indexer__InfoRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "后端查询结果（下游错误也可能以 200 + 业务错误体返回，见各服务说明）。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    true,
                    {
                      "$ref": "#/components/schemas/State__InfoResponse"
                    },
                    {
                      "$ref": "#/components/schemas/Indexer__InfoResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "未知 type 或参数错误（网关直接返回，不进后端）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gateway__Error"
                }
              }
            }
          },
          "404": {
            "description": "后端返回的未知类型（透传）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gateway__Error"
                }
              }
            }
          },
          "501": {
            "description": "后端未实现（透传）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gateway__Error"
                }
              }
            }
          },
          "503": {
            "description": "后端未就绪或未配置（网关直接返回）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gateway__Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Exchange__SignedAction": {
        "type": "object",
        "required": [
          "action",
          "nonce",
          "signature"
        ],
        "properties": {
          "action": {
            "type": "object",
            "description": "exchange action，必须包含 type。L1-agent 签名动作：order/cancel/cancelByCloid/cancelAll/updateLeverage/batchModify；User-signed EIP-712 动作：usdSend/withdraw3/approveAgent。action 字段和值须满足原生 action 格式，JSON 键顺序不影响验签。",
            "oneOf": [
              {
                "$ref": "#/components/schemas/Exchange__OrderAction"
              },
              {
                "$ref": "#/components/schemas/Exchange__CancelAction"
              },
              {
                "$ref": "#/components/schemas/Exchange__CancelByCloidAction"
              },
              {
                "$ref": "#/components/schemas/Exchange__CancelAllAction"
              },
              {
                "$ref": "#/components/schemas/Exchange__UpdateLeverageAction"
              },
              {
                "$ref": "#/components/schemas/Exchange__BatchModifyAction"
              },
              {
                "$ref": "#/components/schemas/Exchange__UsdSendAction"
              },
              {
                "$ref": "#/components/schemas/Exchange__Withdraw3Action"
              },
              {
                "$ref": "#/components/schemas/Exchange__ApproveAgentAction"
              }
            ],
            "discriminator": {
              "propertyName": "type"
            }
          },
          "nonce": {
            "type": "integer",
            "minimum": 1,
            "description": "用户 nonce，必须大于 0。"
          },
          "signature": {
            "$ref": "#/components/schemas/Exchange__SignatureWire"
          },
          "vaultAddress": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{40}$",
            "description": "可选，20 字节 0x hex；存在时参与签名 payload。"
          },
          "expiresAfter": {
            "type": "integer",
            "minimum": 1,
            "description": "可选过期时间；存在时参与签名 payload。"
          }
        }
      },
      "Exchange__SignatureWire": {
        "type": "object",
        "required": [
          "r",
          "s",
          "v"
        ],
        "properties": {
          "r": {
            "type": "string",
            "description": "签名标量 r，SDK 可给变长整数，服务端会左补齐到 32 字节；文档示例按 32 字节 0x hex。"
          },
          "s": {
            "type": "string",
            "description": "签名标量 s，同 r。"
          },
          "v": {
            "type": "integer",
            "enum": [
              27,
              28
            ]
          }
        }
      },
      "Exchange__OrderAction": {
        "type": "object",
        "required": [
          "type",
          "orders",
          "grouping"
        ],
        "properties": {
          "type": {
            "const": "order"
          },
          "orders": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/Exchange__LimitOrder"
            }
          },
          "grouping": {
            "type": "string",
            "example": "na"
          }
        }
      },
...
      "Exchange__CancelAction": {
        "type": "object",
        "required": [
          "type",
          "cancels"
        ],
        "properties": {
          "type": {
            "const": "cancel"
          },
          "cancels": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "a",
                "o"
              ],
              "properties": {
                "a": {
                  "type": "integer"
                },
                "o": {
                  "type": "integer",
                  "description": "链上 oid。"
                }
              }
            }
          }
        }
      },
      "Exchange__CancelByCloidAction": {
        "type": "object",
        "required": [
          "type",
          "cancels"
        ],
        "properties": {
          "type": {
            "const": "cancelByCloid"
          },
          "cancels": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "asset",
                "cloid"
              ],
              "properties": {
                "asset": {
                  "type": "integer"
                },
                "cloid": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "Exchange__CancelAllAction": {
        "type": "object",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "const": "cancelAll"
          }
        }
      },
      "Exchange__UpdateLeverageAction": {
        "type": "object",
        "required": [
          "type",
          "asset",
          "isCross",
          "leverage"
        ],
        "properties": {
          "type": {
            "const": "updateLeverage"
          },
          "asset": {
            "type": "integer"
          },
          "isCross": {
            "type": "boolean"
          },
          "leverage": {
            "type": "integer"
          }
        }
      },
      "Exchange__BatchModifyAction": {
        "type": "object",
        "required": [
          "type",
          "modifies"
        ],
        "properties": {
          "type": {
            "const": "batchModify"
          },
          "modifies": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "oid",
                "order"
              ],
              "properties": {
                "oid": {
                  "type": "integer"
                },
                "order": {
                  "$ref": "#/components/schemas/Exchange__LimitOrder"
                }
              }
            }
          }
        }
      },
      "Exchange__UsdSendAction": {
        "type": "object",
        "description": "User-signed EIP-712：HyperliquidTransaction:UsdSend。nonce 必须等于 action.time。",
        "required": [
          "type",
          "hyperliquidChain",
          "signatureChainId",
          "destination",
          "amount",
          "time"
        ],
        "properties": {
          "type": {
            "const": "usdSend"
          },
          "hyperliquidChain": {
            "type": "string"
          },
          "signatureChainId": {
            "type": "string"
          },
          "destination": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "time": {
            "type": "integer"
          }
        }
      },
      "Exchange__Withdraw3Action": {
        "type": "object",
        "description": "User-signed EIP-712：HyperliquidTransaction:Withdraw。nonce 必须等于 action.time。",
        "required": [
          "type",
          "hyperliquidChain",
          "signatureChainId",
          "destination",
          "amount",
          "time"
        ],
        "properties": {
          "type": {
            "const": "withdraw3"
          },
          "hyperliquidChain": {
            "type": "string"
          },
          "signatureChainId": {
            "type": "string"
          },
          "destination": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "time": {
            "type": "integer"
          }
        }
      },
      "Exchange__ApproveAgentAction": {
        "type": "object",
        "description": "User-signed EIP-712：HyperliquidTransaction:ApproveAgent。外层 nonce 必须等于 action.nonce。",
        "required": [
          "type",
          "hyperliquidChain",
          "signatureChainId",
          "agentAddress",
          "nonce"
        ],
        "properties": {
          "type": {
            "const": "approveAgent"
          },
          "hyperliquidChain": {
            "type": "string"
          },
          "signatureChainId": {
            "type": "string"
          },
          "agentAddress": {
            "type": "string"
          },
          "agentName": {
            "type": "string"
          },
          "nonce": {
            "type": "integer"
          }
        }
      },
      "Exchange__ExchangeResult": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "enum": [
              0,
              1001,
              1002,
              1003
            ],
            "description": "0=accepted（已入本地交易池）；1001 参数/哈希失败；1002 重复交易；1003 签名失败。"
          },
          "message": {
            "type": "string",
            "enum": [
              "accepted",
              "invalid parameter",
              "duplicate transaction",
              "invalid signature"
            ]
          },
          "tx_hash": {
            "type": "string",
            "description": "成功时返回，Keccak256(JSON(SignedAction))，0x 64 位 hex。失败时省略。"
          }
        }
      },
      "Exchange__InfoRequest": {
        "type": "object",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "health",
              "stateInfo",
              "block",
              "bridgeSnapshot",
              "bridgeDepositStatus",
              "bridgeWithdrawalStatus",
              "clearinghouseState",
              "accountOverview",
              "unifiedBalances",
              "accountNonces"
            ],
            "description": "health 由本服务直接回答；其余透传 connected node，block 需带 height。"
          },
          "height": {
            "type": "integer",
            "description": "type=block 时必填。"
          }
        }
      },
      "Exchange__Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "State__InfoRequest": {
        "description": "14 种已实现 type 的联合，discriminator 为 type。",
        "oneOf": [
          {
            "$ref": "#/components/schemas/State__MetaRequest"
          },
          {
            "$ref": "#/components/schemas/State__MetaAndAssetCtxsRequest"
          },
          {
            "$ref": "#/components/schemas/State__AllMidsRequest"
          },
          {
            "$ref": "#/components/schemas/State__L2BookRequest"
          },
          {
            "$ref": "#/components/schemas/State__ClearinghouseStateRequest"
          },
          {
            "$ref": "#/components/schemas/State__OpenOrdersRequest"
          },
          {
            "$ref": "#/components/schemas/State__FrontendOpenOrdersRequest"
          },
          {
            "$ref": "#/components/schemas/State__ExtraAgentsRequest"
          },
          {
            "$ref": "#/components/schemas/State__ActiveAssetDataRequest"
          },
          {
            "$ref": "#/components/schemas/State__UserFeesRequest"
          },
          {
            "$ref": "#/components/schemas/State__RecentTradesRequest"
          },
          {
            "$ref": "#/components/schemas/State__ExchangeStatusRequest"
          },
          {
            "$ref": "#/components/schemas/State__OrderStatusRequest"
          },
          {
            "$ref": "#/components/schemas/State__WebData2Request"
          }
        ],
        "discriminator": {
          "propertyName": "type"
        }
      },
      "State__MetaRequest": {
        "type": "object",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "const": "meta"
          },
          "dex": {
            "type": "string",
            "description": "可选，默认 dex；非默认 dex 返回 501。"
          }
        }
      },
      "State__MetaAndAssetCtxsRequest": {
        "type": "object",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "const": "metaAndAssetCtxs"
          },
          "dex": {
            "type": "string"
          }
        }
      },
      "State__AllMidsRequest": {
        "type": "object",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "const": "allMids"
          },
          "dex": {
            "type": "string"
          }
        }
      },
      "State__L2BookRequest": {
        "type": "object",
        "required": [
          "type",
          "coin"
        ],
        "properties": {
          "type": {
            "const": "l2Book"
          },
          "coin": {
            "type": "string",
            "example": "BTCUSDC"
          },
          "nSigFigs": {
            "type": "integer",
            "description": "可选有效数字位数。"
          },
          "mantissa": {
            "type": "integer",
            "description": "可选尾数。"
          }
        }
      },
      "State__ClearinghouseStateRequest": {
        "type": "object",
        "required": [
          "type",
          "user"
        ],
        "properties": {
          "type": {
            "const": "clearinghouseState"
          },
          "user": {
            "type": "string",
            "description": "42 字符 0x 地址。"
          },
          "dex": {
            "type": "string"
          }
        }
      },
      "State__OpenOrdersRequest": {
        "type": "object",
        "required": [
          "type",
          "user"
        ],
        "properties": {
          "type": {
            "const": "openOrders"
          },
          "user": {
            "type": "string"
          },
          "dex": {
            "type": "string"
          }
        }
      },
      "State__FrontendOpenOrdersRequest": {
        "type": "object",
        "required": [
          "type",
          "user"
        ],
        "properties": {
          "type": {
            "const": "frontendOpenOrders"
          },
          "user": {
            "type": "string"
          },
          "dex": {
            "type": "string"
          }
        }
      },
      "State__ExtraAgentsRequest": {
        "type": "object",
        "required": [
          "type",
          "user"
        ],
        "properties": {
          "type": {
            "const": "extraAgents"
          },
          "user": {
            "type": "string"
          }
        }
      },
      "State__ActiveAssetDataRequest": {
        "type": "object",
        "required": [
          "type",
          "user",
          "coin"
        ],
        "properties": {
          "type": {
            "const": "activeAssetData"
          },
          "user": {
            "type": "string"
          },
          "coin": {
            "type": "string"
          }
        }
      },
      "State__UserFeesRequest": {
        "type": "object",
        "required": [
          "type",
          "user"
        ],
        "properties": {
          "type": {
            "const": "userFees"
          },
          "user": {
            "type": "string"
          }
        }
      },
      "State__RecentTradesRequest": {
        "type": "object",
        "required": [
          "type",
          "coin"
        ],
        "properties": {
          "type": {
            "const": "recentTrades"
          },
          "coin": {
            "type": "string"
          }
        }
      },
...
      "State__InfoResponse": {
        "description": "响应形状随请求 type 变化：meta→State__MetaResponse；metaAndAssetCtxs→[meta, ctxs] 二元数组；allMids→{coin: mid} 对象；l2Book→State__L2BookResponse；clearinghouseState→State__ClearinghouseStateResponse；openOrders→State__OpenOrderResponse 数组；frontendOpenOrders→State__NodeDataOrder 数组；extraAgents→State__ExtraAgentResponse 数组；activeAssetData→State__ActiveAssetDataResponse；userFees→State__UserFeesResponse；recentTrades→State__RecentTradeResponse 数组；exchangeStatus→State__ExchangeStatusResponse；orderStatus→State__OrderStatusResponse；webData2→State__WebData2Response。",
        "oneOf": [
          {
            "$ref": "#/components/schemas/State__MetaResponse"
          },
          {
            "type": "array",
            "prefixItems": [
              {
                "$ref": "#/components/schemas/State__MetaResponse"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/State__PerpAssetCtxResponse"
                }
              }
            ],
            "description": "metaAndAssetCtxs：[meta, 行情数组]，与 universe 一一对应。"
          },
          {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "allMids：{coin: 中间价}，数值均为字符串小数。"
          },
          {
            "$ref": "#/components/schemas/State__L2BookResponse"
          },
          {
            "$ref": "#/components/schemas/State__ClearinghouseStateResponse"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/State__OpenOrderResponse"
            }
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/State__NodeDataOrder"
            }
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/State__ExtraAgentResponse"
            }
          },
          {
            "$ref": "#/components/schemas/State__ActiveAssetDataResponse"
          },
          {
            "$ref": "#/components/schemas/State__UserFeesResponse"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/State__RecentTradeResponse"
            }
          },
          {
            "$ref": "#/components/schemas/State__ExchangeStatusResponse"
          },
          {
            "$ref": "#/components/schemas/State__OrderStatusResponse"
          },
          {
            "$ref": "#/components/schemas/State__WebData2Response"
          }
        ]
      },
...
      "Indexer__InfoRequest": {
        "description": "13 种本地类型联合，discriminator 为 type。",
        "oneOf": [
          {
            "$ref": "#/components/schemas/Indexer__MetaAndAssetCtxsRequest"
          },
          {
            "$ref": "#/components/schemas/Indexer__WebData2Request"
          },
          {
            "$ref": "#/components/schemas/Indexer__AllMidsRequest"
          },
          {
            "$ref": "#/components/schemas/Indexer__CandleSnapshotRequest"
          },
          {
            "$ref": "#/components/schemas/Indexer__OrderStatusRequest"
          },
          {
            "$ref": "#/components/schemas/Indexer__L2BookRequest"
          },
          {
            "$ref": "#/components/schemas/Indexer__FundingHistoryRequest"
          },
          {
            "$ref": "#/components/schemas/Indexer__UserFillsByTimeRequest"
          },
          {
            "$ref": "#/components/schemas/Indexer__UserFillsRequest"
          },
          {
            "$ref": "#/components/schemas/Indexer__RecentTradesRequest"
          },
          {
            "$ref": "#/components/schemas/Indexer__UserFundingRequest"
          },
          {
            "$ref": "#/components/schemas/Indexer__UserNonFundingLedgerUpdatesRequest"
          },
          {
            "$ref": "#/components/schemas/Indexer__HistoricalOrdersRequest"
          }
        ],
        "discriminator": {
          "propertyName": "type"
        }
      },
...
      "Indexer__InfoResponse": {
        "description": "响应形状随请求 type 变化（均为 JSON，数值一律字符串小数，地址恒小写）：metaAndAssetCtxs→[meta, ctxs] 二元数组；webData2→页面聚合对象；allMids→{mids:{coin:px}}对象；candleSnapshot→K 线数组（SDK candle wire {t,T,s,i,o,c,h,l,v,n}）；orderStatus→State__OrderStatusResponse 同形 {status,order?,statusTimestamp?}；l2Book→盘口快照；fundingHistory→资金费数组；userFillsByTime/userFills→Fill 数组（SDK Fill 原文透传）；recentTrades→公开成交数组；userFunding→资金费数组；userNonFundingLedgerUpdates→账本变更数组；historicalOrders→Indexer__HistoricalOrderItem 数组（history.rs 按 HL 风格取最近上限）。下面两个 schema 只给历史订单与订单状态精确建模，其余以本描述为准，后续按需补齐。"
      },
...
      "Gateway__Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
...
}